fix(export to PDF): Notifications now display and hide correctly
