+++ locales/en.json
@@ -37,7 +37,8 @@
   "customReportFooter": "This icon indicates that this CVE has an associated security rule written by Red Hat. See the Vulnerability application for more details.",
   "customReportIntroductionText": "This is a summary of CVEs identified by Red Hat that may impact your Red Hat Enterprise Linux (RHEL) systems.",
   "customReportNoCves": "No CVEs with selected criteria.",
-  "customReportNotification": "Report by CVEs is being generated",
+  "customReportNotificationCve": "Report by CVEs is being generated",
+  "customReportNotificationSystem": "Report by systems is being generated",
   "customReportSearchTermCvssAndPublishDate": "This report includes CVEs {hasSearchTerm, select, yes {matching the search term <b>\"{searchTerm}\"</b>;} no {}} with a CVSS base score of <b>{score}</b>; published {published}",
   "customReportSecurityRuleCheckbox": "Security rule indicator",
   "customReportSystemsExposed": "Count of exposed systems",
