+++ src/Components/SmartComponents/Reports/DownloadCVEsReport.js
@@ -1,5 +1,5 @@
 
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import propTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 import { DownloadButton, CSAwIcon } from '@redhat-cloud-services/frontend-components-pdf-generator';
@@ -11,6 +11,8 @@ import DownloadReport from '../../../Helpers/DownloadReport';
 import { PDF_REPORT_PER_PAGE } from '../../../Helpers/constants';
 import { Text, View } from '@react-pdf/renderer';
 import styles from './Common/styles';
+import { useDispatch } from 'react-redux';
+import { addNotification, clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/cjs/actions';
 
 const DownloadCVEsReport = ({
     filters,
@@ -25,6 +27,13 @@ const DownloadCVEsReport = ({
     const intl = useIntl();
     const [loading, setLoading] = useState(false);
     const [securityRulesPresent, setSecurityRulesPresent] = useState(false);
+    const dispatch = useDispatch();
+
+    useEffect(() => {
+        return () => {
+            dispatch(clearNotifications());
+        };
+    }, [dispatch]);
 
     const otherPagesPDF = data =>
         data.reduce((resultArray, item, index) => {
@@ -37,6 +46,12 @@ const DownloadCVEsReport = ({
 
     const cvesFetch = async () => {
         setLoading(true);
+
+        dispatch(addNotification({
+            dismissable: true,
+            variant: 'info',
+            title: intl.formatMessage(messages.customReportNotificationCve)
+        }));
 
         let { data, meta } = await getCveListByAccount({
             ...params,
@@ -77,6 +92,12 @@ const DownloadCVEsReport = ({
 
     const systemsFetch = async () => {
         setLoading(true);
+
+        dispatch(addNotification({
+            dismissable: true,
+            variant: 'info',
+            title: intl.formatMessage(messages.customReportNotificationSystem)
+        }));
 
         let { data, meta } = await getSystems({
             ...params,
