+++ src/Components/SmartComponents/Reports/DownloadExecutive.js
@@ -1,5 +1,5 @@
 
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useIntl } from 'react-intl';
 import { DownloadButton } from '@redhat-cloud-services/frontend-components-pdf-generator';
 import { getExecutiveReport } from '../../../Helpers/APIHelper';
@@ -12,12 +12,18 @@ import { addNotification, clearNotifications } from '@redhat-cloud-services/fron
 
 const DownloadExecutive = () => {
     const intl = useIntl();
-    const [loading, setLoading] = useState(false);
+    const [isLoading, setLoading] = useState(false);
     const [renderPDF, setRenderPDF] = useState(false);
     const dispatch = useDispatch();
 
     const dataFetch = async () => {
         setLoading(true);
+
+        dispatch(addNotification({
+            dismissable: true,
+            variant: 'info',
+            title: intl.formatMessage(messages.downloadExecutiveNotification)
+        }));
 
         const data = await getExecutiveReport();
         const report = buildExecReport({ data, intl });
@@ -30,18 +36,7 @@ const DownloadExecutive = () => {
 
     const handleDownloadButton = () => {
         setRenderPDF(true);
-        dispatch(addNotification({
-            dismissable: true,
-            variant: 'info',
-            title: intl.formatMessage(messages.downloadExecutiveNotification)
-        }));
     };
-
-    useEffect(() => {
-        return () => {
-            dispatch(clearNotifications());
-        };
-    }, [dispatch]);
 
     return (<Fragment>
         <a href onClick={() => handleDownloadButton()}>
@@ -49,13 +44,16 @@ const DownloadExecutive = () => {
         </a>
         { renderPDF && <DownloadButton
             showButton={false}
-            label={loading ? intl.formatMessage(messages.loading) : intl.formatMessage(messages.executiveReportCardButton)}
+            label={isLoading ? intl.formatMessage(messages.loading) : intl.formatMessage(messages.executiveReportCardButton)}
             asyncFunction={dataFetch}
             buttonProps={{ variant: 'link', isInline: true }}
             type={intl.formatMessage(messages.vulnerabilitiesHeader)}
             fileName={`vulnerability_executive-report--${date}.pdf`}
             allPagesHaveTitle={false}
-            onSuccess={() => setRenderPDF(false)}
+            onSuccess={() => {
+                dispatch(clearNotifications());
+                setRenderPDF(false);
+            }}
         />
         }
     </Fragment>
