+++ src/Components/SmartComponents/Reports/ReportsPage.js
@@ -1,7 +1,6 @@
-import React,  { useState, useEffect } from 'react';
+import React,  { useState } from 'react';
 import { Page, Grid, GridItem, Card, CardTitle, CardBody, CardFooter } from '@patternfly/react-core';
 import { Main } from '@redhat-cloud-services/frontend-components';
-import { useDispatch } from 'react-redux';
 import { FileAltIcon } from '@patternfly/react-icons';
 import DownloadExecutive from './DownloadExecutive';
 import ReportConfigModal from '../Modals/ReportConfigModal';
@@ -11,7 +10,6 @@ import { ChartPieSolid } from '../../PresentationalComponents/CustomIcons/Custom
 import Header from '../../PresentationalComponents/Header/Header';
 import DownloadCVEsReport from '../Reports/DownloadCVEsReport';
 import { constructFilterParameters, buildFilters } from '../Reports/ReportsHelper';
-import { addNotification, clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/cjs/actions';
 import { CVE_REPORT_FILTERS, DEFAULT_FILTER_DATA, PDF_REPORT_USER_NOTE_MAX_LENGTH } from '../../../Helpers/constants';
 import styles from './Common/styles';
 
@@ -22,7 +20,6 @@ const ReportsPage = () => {
     const [userNotes, setUserNotes] = useState('');
     const [columnsToInclude, setColumnsToInclude] = useState(Object.keys(CVE_REPORT_FILTERS));
     const [filterData, setFilterData] = useState(DEFAULT_FILTER_DATA);
-    const dispatch = useDispatch();
 
     const handleModalClose = () => {
         setReportTitle(intl.formatMessage(messages.customReportDefaultTitle));
@@ -35,20 +32,8 @@ const ReportsPage = () => {
 
     const handleDownloadButton = () => {
         setDownloadCVEReport(true);
-        dispatch(addNotification({
-            dismissable: true,
-            variant: 'info',
-            title: intl.formatMessage(messages.customReportNotification)
-        }));
-
         setModalOpen(false);
     };
-
-    useEffect(() => {
-        return () => {
-            dispatch(clearNotifications());
-        };
-    }, [dispatch]);
 
     return (
         <Page>
