+++ src/Components/SmartComponents/SystemsPage/SystemsTableToolbar.js
@@ -103,6 +103,7 @@ const SystemsTableToolbar = ({ selectedHosts, intl, parameters, systems, methods
                 showButton={false}
                 params={parameters}
                 filters={buildActiveFilters(parameters)}
+                onSuccess={() => setExportPDF(false)}
             />
         }
