+++ src/Messages.js
@@ -1307,10 +1307,15 @@ export default defineMessages({
         description: 'Header title for the card with custom executive report',
         defaultMessage: 'Report by CVEs'
     },
-    customReportNotification: {
-        id: 'customReportNotification',
+    customReportNotificationCve: {
+        id: 'customReportNotificationCve',
         description: 'Notification',
         defaultMessage: 'Report by CVEs is being generated'
+    },
+    customReportNotificationSystem: {
+        id: 'customReportNotificationSystem',
+        description: 'Notification',
+        defaultMessage: 'Report by systems is being generated'
     },
     customReportCardDescription: {
         id: 'customReportCardDescription',
